werkfile/requirements.txt
# werkfile/requirements.txt - Python dependencies
Flask
python-dotenv
python-kinde-sdk
xata-py
PyYAML
